后端重构-随机画廊-完成

+++ src/main/java/com/example/randomGallery/service/Impl/PicServiceApiImpl.java
@@ -1,12 +1,13 @@
 package com.example.randomGallery.service.Impl;
+
 import com.example.randomGallery.entity.VO.PicVO;
 import com.example.randomGallery.service.CacheService;
 import com.example.randomGallery.service.PicServiceApi;
 import com.example.randomGallery.service.mapper.PicServiceMapper;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-
 
 
 /**
@@ -21,7 +22,7 @@ public class PicServiceApiImpl implements PicServiceApi {
     private final CacheService cacheService;
 
     @Override
-    public PicVO getInfoById(Integer id) {
+    public PicVO getInfoById(Long id) {
         return picServiceMapper.getInfoById(cacheService.getPicSqlName(), id);
     }
 }
