+++ src/main/java/com/example/randomGallery/service/PicServiceApi.java
@@ -1,11 +1,10 @@
 package com.example.randomGallery.service;
 
-import com.example.randomGallery.entity.VO.GroupVO;
 import com.example.randomGallery.entity.VO.PicVO;
 import org.springframework.stereotype.Service;
 
 @Service
 public interface PicServiceApi {
 
-    PicVO getInfoById(Integer id);
+    PicVO getInfoById(Long id);
 }
